+++ packages/forms/test/components/Form/CheckBoxController.test.tsx
@@ -1,59 +1,60 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import BaseCheckBoxController from '@airbnb/lunar/lib/components/CheckBoxController';
-import { unwrapHOCs } from '@airbnb/lunar-test-utils';
 import CheckBoxController from '../../../src/components/Form/CheckBoxController';
 import { toString } from '../../../src/helpers';
+import { Context } from '../../../src/types';
+import { WrappingFormComponent, createFormContext } from '../../utils';
 
 describe('<CheckBoxController />', () => {
-  const form = {
-    change() {},
-    getState: () => ({} as any),
-    register: jest.fn(),
-  };
+  let context: Context;
+
+  beforeEach(() => {
+    context = createFormContext();
+  });
 
   it('connects to the form', () => {
-    const wrapper = unwrapHOCs(
-      shallow(
-        <CheckBoxController label="Label" name="foo" defaultValue="bar" validator={() => {}}>
-          {CB => (
-            <div>
-              <CB value="foo" label="Foo" />
-              <CB value="bar" label="Bar" />
-              <CB value="baz" label="Baz" />
-            </div>
-          )}
-        </CheckBoxController>,
-      ),
-      'FormCheckBoxController',
-      form,
+    const wrapper = mount(
+      <CheckBoxController label="Label" name="foo" defaultValue="bar" validator={() => {}}>
+        {CB => (
+          <div>
+            <CB value="foo" label="Foo" />
+            <CB value="bar" label="Bar" />
+            <CB value="baz" label="Baz" />
+          </div>
+        )}
+      </CheckBoxController>,
+      {
+        wrappingComponent: WrappingFormComponent,
+        wrappingComponentProps: { context },
+      },
     );
 
-    expect(form.register).toHaveBeenCalledWith(
+    expect(context.register).toHaveBeenCalledWith(
       expect.objectContaining({ name: 'foo', defaultValue: 'bar', parse: toString }),
       expect.anything(),
     );
 
-    expect(wrapper.shallow().find(BaseCheckBoxController)).toHaveLength(1);
+    expect(wrapper.find(BaseCheckBoxController)).toHaveLength(1);
   });
 
   it('passes value as an array', () => {
-    const wrapper = unwrapHOCs(
-      shallow(
-        <CheckBoxController label="Label" name="foo" defaultValue="bar" validator={() => {}}>
-          {CB => (
-            <div>
-              <CB value="foo" label="Foo" />
-              <CB value="bar" label="Bar" />
-              <CB value="baz" label="Baz" />
-            </div>
-          )}
-        </CheckBoxController>,
-      ),
-      'FormCheckBoxController',
-      form,
+    const wrapper = mount(
+      <CheckBoxController label="Label" name="foo" defaultValue="bar" validator={() => {}}>
+        {CB => (
+          <div>
+            <CB value="foo" label="Foo" />
+            <CB value="bar" label="Bar" />
+            <CB value="baz" label="Baz" />
+          </div>
+        )}
+      </CheckBoxController>,
+      {
+        wrappingComponent: WrappingFormComponent,
+        wrappingComponentProps: { context },
+      },
     );
 
-    expect(wrapper.prop('value')).toEqual(['bar']);
+    expect(wrapper.find(BaseCheckBoxController).prop('value')).toEqual(['bar']);
   });
 });