+++ packages/forms/test/components/Form/FileInput.test.tsx
@@ -1,28 +1,31 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import BaseFileInput from '@airbnb/lunar/lib/components/FileInput';
-import { unwrapHOCs } from '@airbnb/lunar-test-utils';
 import FileInput from '../../../src/components/Form/FileInput';
+import { Context } from '../../../src/types';
+import { WrappingFormComponent, createFormContext } from '../../utils';
 
 describe('<FileInput />', () => {
-  const form = {
-    change() {},
-    getState: () => ({} as any),
-    register: jest.fn(),
-  };
+  let context: Context;
+
+  beforeEach(() => {
+    context = createFormContext();
+  });
 
   it('connects to the form', () => {
-    const wrapper = unwrapHOCs(
-      shallow(<FileInput label="Label" name="foo" defaultValue="bar" validator={() => {}} />),
-      'FormFileInput',
-      form,
+    const wrapper = mount(
+      <FileInput label="Label" name="foo" defaultValue="bar" validator={() => {}} />,
+      {
+        wrappingComponent: WrappingFormComponent,
+        wrappingComponentProps: { context },
+      },
     );
 
-    expect(form.register).toHaveBeenCalledWith(
+    expect(context.register).toHaveBeenCalledWith(
       expect.objectContaining({ name: 'foo', defaultValue: 'bar' }),
       expect.anything(),
     );
 
-    expect(wrapper.shallow().find(BaseFileInput)).toHaveLength(1);
+    expect(wrapper.find(BaseFileInput)).toHaveLength(1);
   });
 });