+++ packages/core/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 2.13.1 - 2019-07-25
+
+#### 🎨 Styles
+
+- **[Button,Input,Autocomplete,Menu]** Normalize and standardize spacing (#113) ([9151b3e](https://github.com/airbnb/lunar/tree/master/packages/core/commit/9151b3e)), closes [#113](https://github.com/airbnb/lunar/tree/master/packages/core/issues/113)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
+## 2.13.0 - 2019-07-24
+
+#### 🚀 Updates
+
+- Refine and use explicit value types. Allow null and undefined default values. (#132) ([ecede51](https://github.com/airbnb/lunar/tree/master/packages/core/commit/ecede51)), closes [#132](https://github.com/airbnb/lunar/tree/master/packages/core/issues/132)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
+### 2.12.1 - 2019-07-24
+
+#### 🐞 Fixes
+
+- **[Tooltip]** Fix cannot read property getBoundingClientRect of null. (#133) ([196582b](https://github.com/airbnb/lunar/tree/master/packages/core/commit/196582b)), closes [#133](https://github.com/airbnb/lunar/tree/master/packages/core/issues/133)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
 ## 2.12.0 - 2019-07-23
 
 #### 🚀 Updates